requirements.txt
flask
mysql-connector-python
google-generativeai
werkzeug
